Fix more warnings

+++ Baikal/Utils/cl_program.cpp
@@ -190,20 +190,20 @@ CLWProgram CLProgram::Compile(const std::string &opts)
         file.close();
 #endif
     }
-    catch (CLWException exception)
-    {
-        std::cerr<<"Compilation failed!"<<std::endl;
-        std::cerr<<"Dumping source to file:"<<m_program_name<<".cl.failed"<<std::endl;
-        std::string fname = m_program_name + ".cl.failed";
-        std::ofstream file(fname);
-        file << m_compiled_source;
-        file.close();
-        throw;
-    }
+	catch (CLWException exception)
+	{
+		std::cerr << "Compilation failed!" << std::endl;
+		std::cerr << "Dumping source to file:" << m_program_name << ".cl.failed" << std::endl;
+		std::string fname = m_program_name + ".cl.failed";
+		std::ofstream file(fname);
+		file << m_compiled_source;
+		file.close();
+		throw;
+	}
 
-    end = std::chrono::high_resolution_clock::now();
-    int elapsed_ms = std::chrono::duration_cast<std::chrono::milliseconds>(end-start).count();
-    std::cerr<<"Program compilation time: "<<elapsed_ms<<" ms"<<std::endl;
+	end = std::chrono::high_resolution_clock::now();
+	int elapsed_ms = (int)std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
+	std::cerr << "Program compilation time: " << elapsed_ms << " ms" << std::endl;
 
     m_is_dirty = false;
     return compiled_program;
